authentication flow working
